nacl-mozc, chromeos-hangul: Fix arm64 installs
Keep nacl files for arm64 builds.

BUG=b:188495414
TEST=CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4009410
Auto-Submit: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>

+++ app-i18n/chromeos-hangul/chromeos-hangul-1.2.0.7-r2.ebuild
@@ -14,7 +14,7 @@ KEYWORDS="*"
 
 src_prepare() {
 	# Removes unused NaCl binaries.
-	if ! use arm ; then
+	if ! use arm && ! use arm64; then
 		rm hangul_arm.nexe || die
 	fi
 	if ! use x86 ; then
